Create unl-report.json locale id
## Proposed Change

### Description

This proposed change introduces translation updates for the JSON script to provide a more user-friendly experience for Indonesian speakers. The translation includes content related to the UNL Report.

### Context

This change aims to enhance the usability of features related to the UNL Report by providing accurate and clear information in Indonesian.